fixed rotation + orientation

+++ inc/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: cbouvet <cbouvet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/13 17:02:09 by cbouvet           #+#    #+#             */
-/*   Updated: 2024/06/05 21:13:39 by cbouvet          ###   ########.fr       */
+/*   Updated: 2024/06/05 21:58:05 by cbouvet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -198,6 +198,6 @@ void	cub3d_maker(t_player *player);
 void	set_variables(t_player *player, double cam);
 int	step_incr(t_player *player, int axis);
 double	raycaster(t_player *player, int step_x, int step_y);
-void	visual_render(t_player *player, int i, double wall_height);
+void	visual_render(t_player *player, int i, int wall_height);
 
 #endif
